src/models/basemodel.py
from sqlalchemy import (
    Column, String, Text, ForeignKey,
    DateTime, ARRAY
)
from sqlalchemy.dialects.postgresql import UUID
from sqlalchemy.orm import relationship, declarative_base
from sqlalchemy.sql import func
import uuid

Base = declarative_base()